better channels (rip unfrom)

+++ src/text.rs
@@ -1,5 +1,5 @@
 use serde::{Serialize, Deserialize};
-use midi_msg::{MidiMsg, ChannelVoiceMsg, Channel};
+use midi_msg::{MidiMsg, ChannelVoiceMsg};
 
 use crate::midi::*;
 use crate::graphics::RGB;
@@ -26,8 +26,6 @@ pub enum TextDirection {
 pub struct Text {
     pub xpos: u32,
     pub ypos: u32,
-    #[serde(serialize_with = "unfrom_channel")]
-    #[serde(deserialize_with = "from_channel")]
     pub channel: Channel,
     pub base_note: u8,
     pub direction: TextDirection,
@@ -49,7 +47,7 @@ impl MidiProcessor for Text {
     fn deal_with(&mut self, message: MidiMsg) {
         match message {
             MidiMsg::ChannelVoice {channel, msg} => {
-                if channel == self.channel {
+                if channel as u8 == self.channel as u8 {
                     match msg {
                         ChannelVoiceMsg::NoteOn { note, velocity } => {
                             let n_base = 30;
